[FEATURE] Add MigrateRequiredFlagSiteConfigRector
Resolves: #2944
Releases: 3, 2
(cherry picked from commit 9269cde04139ee781bea20dcc5703abd5a21d869)
