+++ src/Rector/AbstractArrayDimFetchTcaRector.php
@@ -10,12 +10,19 @@ use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Scalar\String_;
 use Rector\PhpParser\Node\Value\ValueResolver;
 use Rector\Rector\AbstractRector;
+use Ssch\TYPO3Rector\Helper\TcaHelperTrait;
 
 /**
  * Base rector that detects Assignments containing TCA definitions and allows to refactor them
  */
 abstract class AbstractArrayDimFetchTcaRector extends AbstractRector
 {
+    use TcaHelperTrait;
+
+    protected const CONFIG = 'config';
+
+    protected bool $hasAstBeenChanged = false;
+
     protected ValueResolver $valueResolver;
 
     public function __construct(ValueResolver $valueResolver)
