+++ tests/Rector/v12/v0/MigrateRequiredFlagSiteConfigRector/config/configured_rule.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Ssch\TYPO3Rector\TYPO312\v0\MigrateRequiredFlagSiteConfigRector;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->import(__DIR__ . '/../../../../../../config/config_test.php');
+    $rectorConfig->rule(MigrateRequiredFlagSiteConfigRector::class);
+};
